refactor: move shape map functions to the ArrowWeb.ShapeView module
